testing if the temp folder exists before clearing it
Former-commit-id: 62e53dcd50ebd9b71761d4e25e5b23d44d178312
Former-commit-id: 36ee337dd6f10f9f133ddcca3788bd7e843bfc1a
Former-commit-id: 8b39cc871c0a09239c73a04cdf0ed3b772f37283
Former-commit-id: 10508d6498bc9eb0fa50eb78fb1858cb46ef7700
Former-commit-id: 47fc8d0988d692f93c490065cc5794c3859c6888
Former-commit-id: 225a1335aeafac8f6d49e8581525ae66f5072b22
Former-commit-id: c9e59c6d161c17e7e025c306ba83f99def9a7c73
Former-commit-id: c04f77a1dce583ed4e9797a8f47436f0c49e30f2

+++ Custom Consoles/PowerShell/Exports/Functions/Customisation Helpers/Clear-PowerShellTempFolder.ps1
@@ -1,7 +1,9 @@
 function Clear-PowerShellTempFolder {
 	try{
 		Push-Location $ProfileConfig.General.UserFolder
-		Remove-Item $ProfileConfig.General.TempFolder -Force -Recurse
+		if (Test-Path $ProfileConfig.General.TempFolder) {
+			Remove-Item $ProfileConfig.General.TempFolder -Force -Recurse
+		}
 		New-Item $ProfileConfig.General.TempFolder -Type Directory -Force | Out-Null
 	}
 	finally {
